Withdraw the direct ensure wake; keep the pre-signal trim
Instrumented CI E2E runs proved the direct wake regresses reply
correctness under slow timing: runtime import/consume watermarks
publish at checkpoint, so the Temporal workflow's reconcile-ensure loop
treats a direct-woken pointer as still-lagging and lands extra ensures
in inter-invocation gaps — four fresh invocations in twelve seconds for
one message, with one raced invocation re-running an already-replied
turn and delivering a duplicate reply (two message posts observed for
one webhook). Deterministic across three CI runs; invisible locally
only because fast hardware closes the window.

The direct wake, dual-auth ensure route, control-client method, and
triggeredByWebDirect leaf are removed. Any future direct wake must
carry pointer awareness into the workflow so it skips its immediate
ensure (documented in ARCHITECTURE.md and hosted-runtime-protocol.md).
The PR keeps the planner checkpoint threading and pre-signal trim, the
after() schedulers on all webhook routes, and the E2E stub request
diagnostics.
