+++ agent-docs/SECURITY.md
@@ -35,7 +35,6 @@ Last verified: 2026-06-28
 - Companion (iOS) device-sync routes under `/api/device-sync/companion/**` authenticate only with a Privy identity token in `Authorization: Bearer` (no cookie fallback, so no browser ambient authority or CSRF surface). The Junction SDK sign-in token they mint is short-lived, returned exactly once, and must never be logged or persisted; the account-ensure step must reuse the shared device-sync `upsertConnection` external-account identity discipline so SDK and Junction Link flows always share one `device_connection`. The junction client's API-key-prefix/environment validation is the sandbox/production separation authority, and the response surfaces the active environment.
 - Direct Strava webhook POST delivery must fail closed unless the provider-owned signing secret verifies the `X-Strava-Signature` timestamp and HMAC over the raw body. The Strava GET verify token is only a subscription-challenge/admin secret, not POST delivery authentication. Junction-backed device sync uses Junction's own webhook signature boundary and must not depend on direct Strava webhook trust.
 - Hosted Linq first-contact admission is a web-owned OpenAI egress path for unknown signup-link candidates only. Keep `HOSTED_ONBOARDING_LINQ_FIRST_CONTACT_ADMISSION_OPENAI_API_KEY` or the fallback `OPENAI_API_KEY` in web environment configuration only. The approved classifier input is bounded first-contact text plus sparse contact-kind/part-type metadata and a fixed `imessage`/`sms`/`rcs`/`unknown` service enum; do not add member ids, raw provider payloads, routing secrets, invite codes, mailbox bodies, transcripts, contact lookup keys, attachments, or prior conversation context. Do not persist classifier prompts, raw responses, model rationales, provider response bodies, or raw first-contact message text. The legacy nullable rejected-message-text column is retained only as an ignored deploy-skew compatibility column during the expand/contract rollout, and the migration scrubs stored values rather than dropping the column under old app code. Persist only the event-id keyed terminal allow/block decision with confidence/source so replay and concurrency observe the same admission result, and keep that decision write duplicate-safe by event id without relying on caught unique-constraint errors inside open transactions. Logs may include only sanitized confidence/source/failure metadata, safe bounded provider error code/type/message/request-id-presence, and event id suffixes. When enforcement is enabled, textless deterministic blocks, explicit classifier blocks, OpenAI refusal, content-filter outcomes, and first-contact budget exhaustion must be acknowledged as blocked without member creation, invite creation, reply send, read receipt, wake, or mailbox side effects. Classifier-unavailable states, including missing keys, timeouts, ordinary non-2xx responses, malformed output, non-completed OpenAI responses, max-output exhaustion, and OpenAI quota or credits exhaustion, intentionally fail open by recording a deterministic allow decision so a legitimate first contact is not permanently dropped.
-- The Cloudflare `runtime/ensure-processing` route accepts exactly two credentials: the Temporal orchestrator's web-callback signature and web's Vercel OIDC identity (the same identity already used by the browser-vault/status/deletion control routes). Authorization dispatches on the credential the caller presented and never falls through a failed signature to OIDC or vice versa. The web direct wake carries no message payload, mints diagnostics-only `web-ingress-` attempt ids, and grants web no authority it did not already exercise through the Temporal signal; the `triggeredByWebDirect` diagnostic is derived from the authorizing credential, never from caller-supplied fields.
 - Cloudflare Durable Object storage that holds hosted gateway projections, gateway event logs, or similar hot derived user data must stay encrypted at rest with the worker's hosted storage crypto rather than being persisted as plaintext coordination state.
 - Hosted R2 storage namespace ids are metadata-opacity helpers, not auth boundaries. The deterministic `hsn_${hash(userId)}` namespace is acceptable only while hosted user ids remain high-entropy random ids, R2 buckets stay private, and stored payloads stay encrypted. Do not log, export, or expose hosted R2 object keys or namespace ids to users or third parties; enforce access through signed control-plane requests, Durable Object user binding, crypto context authority, and explicit ownership checks. If user ids become guessable or any R2 object key/listing can leak outside trusted runtime operators, switch the namespace to a stored random per-user `storageNamespaceId` or a versioned secret-HMAC-derived value before exposing that surface.
 - Hosted per-user env overrides are for per-user credentials and verified-identity metadata only. They must never be allowed to set executable selectors or process-control variables such as `FFMPEG_COMMAND`, `WHISPER_COMMAND`, `WHISPER_MODEL_PATH`, `NODE_OPTIONS`, `LD_PRELOAD`, or `DYLD_INSERT_LIBRARIES`, because those values can steer subprocess execution inside the hosted runner.
