+++ apps/web/test/hosted-execution-handoff.test.ts
@@ -495,7 +495,6 @@ describe("deleteHostedRunnerUserDataBestEffort", () => {
     vi.mocked(readHostedExecutionControlClientIfConfigured).mockReturnValue({
       createBrowserVaultSession: vi.fn(),
       deleteUserData,
-      ensureRuntimeProcessing: vi.fn(),
       getRunnerStatus: vi.fn(),
     } as ReturnType<typeof readHostedExecutionControlClientIfConfigured>);
 
@@ -525,7 +524,6 @@ describe("deleteHostedRunnerUserDataBestEffort", () => {
     vi.mocked(readHostedExecutionControlClientIfConfigured).mockReturnValue({
       createBrowserVaultSession: vi.fn(),
       deleteUserData,
-      ensureRuntimeProcessing: vi.fn(),
       getRunnerStatus: vi.fn(),
     } as ReturnType<typeof readHostedExecutionControlClientIfConfigured>);
